refactor: fix metrics registration

+++ YourBonoPlatform/Bonds/Domain/Model/Aggregates/Bond.cs
@@ -10,23 +10,23 @@ public class Bond
     {
         UserId = 0;
         Name = string.Empty;
-        NominalValue = 0.0;
-        MarketValue = 0.0;
+        NominalValue = 0.0m;
+        MarketValue = 0.0m;
         Duration = 0;
         Frequency = 0;
         InterestRateTypeId = 0;
-        InterestRate = 0.0;
+        InterestRate = 0.0m;
         Capitalization = 0;
-        DiscountRate = 0.0;
+        DiscountRate = 0.0m;
         EmissionDate = DateTime.MinValue;
         GracePeriodTypeId = 0;
         GracePeriodDuration = 0;
         CurrencyTypeId = 0;
-        PrimeRate = 0.0;
-        StructuredRate = 0.0;
-        PlacementRate = 0.0;
-        FloatingRate = 0.0;
-        CavaliRate = 0.0;
+        PrimeRate = 0.0m;
+        StructuredRate = 0.0m;
+        PlacementRate = 0.0m;
+        FloatingRate = 0.0m;
+        CavaliRate = 0.0m;
     }
     public Bond(CreateBondCommand command)
     {
@@ -76,21 +76,21 @@ public class Bond
     public int Id { get; private set; }
     public int UserId { get; private set; }
     public string Name { get; private set; }
-    public double NominalValue { get; private set; }
-    public double MarketValue { get; private set; }
+    public decimal NominalValue { get; private set; }
+    public decimal MarketValue { get; private set; }
     public int Duration { get; private set; }
     public int Frequency { get; private set; }
     public int InterestRateTypeId { get; private set; }
-    public double InterestRate { get; private set; }
+    public decimal InterestRate { get; private set; }
     public int Capitalization { get; private set; }
-    public double DiscountRate { get; private set; }
+    public decimal DiscountRate { get; private set; }
     public DateTime EmissionDate { get; private set; }
     public int GracePeriodTypeId { get; private set; }
     public int GracePeriodDuration { get; private set; }
     public int CurrencyTypeId { get; private set; }
-    public double PrimeRate { get; private set; }
-    public double StructuredRate { get; private set; }
-    public double PlacementRate { get; private set; }
-    public double FloatingRate { get; private set; }
-    public double CavaliRate { get; private set; }
+    public decimal PrimeRate { get; private set; }
+    public decimal StructuredRate { get; private set; }
+    public decimal PlacementRate { get; private set; }
+    public decimal FloatingRate { get; private set; }
+    public decimal CavaliRate { get; private set; }
 }
