+++ YourBonoPlatform/Bonds/Domain/Model/Entities/BondMetrics.cs
@@ -2,7 +2,7 @@ namespace YourBonoPlatform.Bonds.Domain.Model.Entities;
 
 public class BondMetrics
 {
-    public BondMetrics(int id, int bondId, double maxPrice, double duration, double convexity, double modifiedDuration, double tcea, double trea)
+    public BondMetrics(int id, int bondId, decimal maxPrice, decimal duration, decimal convexity, decimal modifiedDuration, decimal tcea, decimal trea)
     {
         Id = id;
         BondId = bondId;
@@ -25,10 +25,10 @@ public class BondMetrics
     
     public int Id { get; private set; }
     public int BondId { get; private set; }
-    public double MaxPrice { get; private set; }
-    public double Duration { get; private set; }
-    public double Convexity { get; private set; }
-    public double ModifiedDuration { get; private set; }
-    public double Tcea { get; private set; }
-    public double Trea { get; private set; }
+    public decimal MaxPrice { get; private set; }
+    public decimal Duration { get; private set; }
+    public decimal Convexity { get; private set; }
+    public decimal ModifiedDuration { get; private set; }
+    public decimal Tcea { get; private set; }
+    public decimal Trea { get; private set; }
 }
