+++ YourBonoPlatform/Bonds/Interfaces/REST/Resources/BondMetricsResource.cs
@@ -3,10 +3,10 @@ namespace YourBonoPlatform.Bonds.Interfaces.REST.Resources;
 public record BondMetricsResource(
     int Id,
     int BondId,
-    double MaxPrice,
-    double Duration,
-    double Convexity,
-    double ModifiedDuration,
-    double Tcea,
-    double Trea
+    decimal MaxPrice,
+    decimal Duration,
+    decimal Convexity,
+    decimal ModifiedDuration,
+    decimal Tcea,
+    decimal Trea
     );
